Annotate get_terrain_name(), so pyright will lint cleanly

+++ tests/bench2.py
@@ -89,7 +89,7 @@ class DeterministicEnvironment2(DeterministicEnvironment):
                     healing_power = (1 - (border_avg / 2)) - dist_penal
                 else:
                     healing_power = -0.5
-                troop.health += healing_power / 25
+                troop.health += int(healing_power / 25)
                 if troop.health > 100:
                     troop.health = 100
